requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-resized==1.0.2
Faker==19.3.1
idna==3.4
Pillow==10.0.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
urllib3==2.0.4
